requirements.txt
Django==1.9.5
Pillow==3.2.0
django-cors-headers==1.1.0
django-crispy-forms==1.6.0
django-dual-authentication==1.0.0
django-easy-pjax==1.3.0
django-fm==0.2.1
djangorestframework==3.3.3
mutagen==1.31
